fix: fixing moving of issues, swapping of tasks, moving issue to another table.

+++ main-service/src/main/kotlin/io/dnajd/mainservice/service/project_table/ProjectTableServiceImpl.kt
@@ -82,7 +82,7 @@ class ProjectTableServiceImpl(
             throw IllegalArgumentException("Tables don't belong to the same project $fId $sId")
         }
 
-        tableRepository.swapPositions(fId.toInt(), sId.toInt())
+        tableRepository.swapPositions(fId, sId)
     }
 
     override fun deleteById(id: Long) {
